+++ src/components/layout/index.ts
@@ -1,2 +1,2 @@
 export { default as Navbar } from './Navbar';
-export { default as Header } from './Header';
+export { default as Header } from './header/Header';
